Mem leak fix

See merge request matf-bg-ac-rs/course-rs/projects-2023-2024/MATFWars!22

+++ MATFWars/ObstacleNode.cpp
@@ -18,7 +18,7 @@ void ObstacleNode::paint(QPainter *painter, const QStyleOptionGraphicsItem *opti
     Q_UNUSED(widget);
 
     QColor brushColor = QColor(Qt::darkCyan);
-    brushColor.setAlpha(255 * m_obstacleNode->health() / m_obstacleNode->maxHealth());
+    brushColor.setAlpha(static_cast<int>(255 * m_obstacleNode->health() / m_obstacleNode->maxHealth()));
 
     painter->setBrush(brushColor);
     painter->setPen(Qt::NoPen);
